requirements.txt
appdirs==1.4.4
attrs==19.3.0
bcrypt==3.1.7
blessed==1.17.5
botocore==1.15.49
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
future==0.16.0
gunicorn==20.0.4
idna==2.7
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.2
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.0.0
pipenv==2018.11.26
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.0
PyYAML==5.3.1
requests==2.20.1
rsa==3.4.2
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
virtualenv==20.0.21
virtualenv-clone==0.5.4
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.0
zipp==3.1.0
